implement security measurements that prevent other users to delete other's content

+++ server.py
@@ -153,7 +153,8 @@ def route_delete_question(question_id):
 
 @app.route('/question/<question_id>/<answer_id>/delete')
 def route_delete_answer(question_id, answer_id):
-    data_manager.delete_answer(answer_id)
+    if data_manager.answer_belongs_to_user(session['username'], answer_id):
+        data_manager.delete_answer(answer_id)
     return redirect(url_for('display_question_and_answers', question_id=question_id))
 
 
@@ -163,11 +164,13 @@ def route_edit_answer(answer_id):
     question_id = answer_data.get('question_id')
     question_data = data_manager.get_single_entry('question', question_id)
 
-    if request.method == 'GET':
-        return render_template('database_ops/new_answer.html', answer=answer_data, question=question_data)
+    if data_manager.answer_belongs_to_user(session['username'], answer_id):
 
-    user_inputs_for_answer = request.form.to_dict()
-    data_manager.update_entry('answer', answer_id, user_inputs_for_answer)
+        if request.method == 'GET':
+            return render_template('database_ops/new_answer.html', answer=answer_data, question=question_data)
+
+        user_inputs_for_answer = request.form.to_dict()
+        data_manager.update_entry('answer', answer_id, user_inputs_for_answer)
 
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
@@ -250,18 +253,19 @@ def route_tags():
 @app.route('/comment/<comment_id>/delete', methods=["GET", "POST"])
 def route_delete_comment(comment_id):
     comment = data_manager.get_single_entry('comment', comment_id)
-    answer_id_of_comment = comment['answer_id']
-    question_id_of_comment = comment['question_id']
+    if data_manager.comment_belongs_to_user(session['username'], comment_id):
+        answer_id_of_comment = comment['answer_id']
+        question_id_of_comment = comment['question_id']
 
-    if request.method == "GET":
-        if answer_id_of_comment:
-            answer_data = data_manager.get_single_entry('answer', answer_id_of_comment)
-            return render_template('database_ops/delete_comment.html', answer=answer_data['message'], comment=comment)
-        question_data = data_manager.get_single_entry('question', question_id_of_comment)
-        return render_template('database_ops/delete_comment.html', question=question_data, comment=comment)
+        if request.method == "GET":
+            if answer_id_of_comment:
+                answer_data = data_manager.get_single_entry('answer', answer_id_of_comment)
+                return render_template('database_ops/delete_comment.html', answer=answer_data['message'], comment=comment)
+            question_data = data_manager.get_single_entry('question', question_id_of_comment)
+            return render_template('database_ops/delete_comment.html', question=question_data, comment=comment)
 
-    if request.form['delete-button'] == 'Yes':
-        data_manager.delete_comment(comment_id)
+        if request.form['delete-button'] == 'Yes':
+            data_manager.delete_comment(comment_id)
 
     return redirect(url_for('display_question_and_answers', question_id=comment['question_id']), code=307)
 
@@ -270,17 +274,19 @@ def route_delete_comment(comment_id):
 def route_edit_comment(comment_id):
     comment_data = data_manager.get_single_entry('comment', comment_id)
     question_id = comment_data.get('question_id')
-    question_data = data_manager.get_single_entry('question', question_id)
-    answer_data = None
-    if comment_data['answer_id']:
-        answer_id = comment_data['answer_id']
-        answer_data = data_manager.get_single_entry('answer', answer_id)
 
-    if request.method == 'GET':
-        return render_template('database_ops/new_comment.html', comment=comment_data, answer=answer_data, question=question_data)
+    if data_manager.comment_belongs_to_user(session['username'], comment_id):
+        question_data = data_manager.get_single_entry('question', question_id)
+        answer_data = None
+        if comment_data['answer_id']:
+            answer_id = comment_data['answer_id']
+            answer_data = data_manager.get_single_entry('answer', answer_id)
 
-    new_comment_message = request.form['message']
-    data_manager.update_comment_message(comment_data, new_comment_message)
+        if request.method == 'GET':
+            return render_template('database_ops/new_comment.html', comment=comment_data, answer=answer_data, question=question_data)
+
+        new_comment_message = request.form['message']
+        data_manager.update_comment_message(comment_data, new_comment_message)
 
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
